requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
eval_type_backport==0.3.1
faiss-cpu==1.9.0
filelock==3.24.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
importlib-metadata==6.11.0
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.63
langchain-mistralai==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.1.147
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mistralai==1.12.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
numpy==1.26.4
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==23.2
pandas==2.1.4
pillow==10.4.0
pluggy==1.6.0
propcache==0.4.1
protobuf==4.25.8
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
streamlit==1.29.0
tenacity==8.5.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.66.1
typer==0.24.0
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
validators==0.35.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
